pre analysis qc controller
