+++ cg/meta/workflow/pre_analysis_quality_check/models.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+class SequencingQualityMetrics(BaseModel):
+    pass
+
+
+class PreAnalysisQualityMetrics(BaseModel):
+    sequencing_metrics: SequencingQualityMetrics = None
